+++ src/activation/noact.rs
@@ -3,11 +3,11 @@ use super::ActivationFunction;
 pub struct NoActivation;
 
 impl ActivationFunction for NoActivation {
-    fn func(x: f64) -> f64 {
+    fn func(x: f32) -> f32 {
         x
     }
 
-    fn deriv(_: f64) -> f64 {
+    fn deriv(_: f32) -> f32 {
         1.
     }
 }
