+++ src/loss/mod.rs
@@ -5,11 +5,11 @@ pub mod mse;
 
 pub trait LossFunction<const N: usize> {
     fn func(
-        y_out: SVector<f64, N>,
-        y_test: SVector<f64, N>,
-    ) -> f64;
+        y_out: SVector<f32, N>,
+        y_test: SVector<f32, N>,
+    ) -> f32;
     fn grad(
-        y_out: SVector<f64, N>,
-        y_test: SVector<f64, N>,
-    ) -> SVector<f64, N>;
+        y_out: SVector<f32, N>,
+        y_test: SVector<f32, N>,
+    ) -> SVector<f32, N>;
 }
