+++ src/optimizers/adagrad.rs
@@ -10,7 +10,7 @@ pub struct Adagrad<
     const R: usize,
     const C: usize,
 > {
-    g: SMatrix<f64, R, C>,
+    g: SMatrix<f32, R, C>,
 }
 
 impl<
@@ -28,10 +28,10 @@ impl<
 
     fn update_param(
         &mut self,
-        weight: &mut SMatrix<f64, R, C>,
-        gradient: &SMatrix<f64, R, C>,
+        weight: &mut SMatrix<f32, R, C>,
+        gradient: &SMatrix<f32, R, C>,
     ) {
-        let alpha = ALPHA_NUM as f64 / ALPHA_DEN as f64;
+        let alpha = ALPHA_NUM as f32 / ALPHA_DEN as f32;
         self.g += gradient.component_mul(gradient);
         *weight -= alpha
             * component_invsqrt(&self.g)
